Module Fmlib_js

Javascript wrappers.

Common Modules

Usable in a browser window, a worker thread in the browser and node.

module Base : sig ... end

Basic functions for the interaction between ocaml and javascript.

module Timer : sig ... end

One shot timers and interval timers

module Date : sig ... end

The javascript date object

module Url : sig ... end

URL utility functions

Common Browser Modules

Usable in a browser window or in a worker thread in the browser.

module Event : sig ... end

Browser events.

module Event_target : sig ... end

Event_targets in the browser

module File : sig ... end

The javascript file object

module File_reader : sig ... end

The javascript file reader object

module Http_request : sig ... end

Http requests made in the browser.

module Web_worker : sig ... end

Create webworker in the main thread and in the worker thread.

Browser Window Modules

module Dom : sig ... end

The content of a browser window.