Module Fmlib_js.Dom

The content of a browser window.

Hierarchy:

    Document                        Text node

    Element                         Character data

                    Node

                    Event Target

I.e. a document is an element which is a node which is an event target. A text node is an event target as well.

Each document has a body element which is the root of the visible document.

Nodes can form a tree. I.e. each node has an optional parent, an optional first child and an optional sibling. There are methods to add and remove children from a node.

Note: Use the code only within a browser window. Never in node and never in a worker thread.

module Node : sig ... end

A node in the dom tree.

module Style : sig ... end

Inline style of a document element.

module Element : sig ... end

Element node.

module Document : sig ... end

module Location : sig ... end

Module representing location.

module Storage : sig ... end

Module representing local or session storage.

module History : sig ... end

Module representing the browser history.

module Window : sig ... end

Module representing a browser window.