Module Dom.Style

Inline style of a document element.

type t

val set : string -> string -> t -> unit

set name value style Set the property name to value in style.

Examples:

    set "background-color" "red"      style
    set "border"           "10px red" style

val remove : string -> t -> unit

remove name style Remove the property name from style.